refactor: 자동차 이동거리 자료형 int 변경

+++ src/main/java/racingcar/car/Car.java
@@ -2,7 +2,7 @@ package racingcar.car;
 
 public abstract class Car {
     private String carName;
-    protected long moveDistance;
+    protected int moveDistance;
 
     public Car() {
     }
